requirements.txt
typer>=0.9.0
rich>=13.3.5
click>=8.1.3
pydantic>=2.0.0

cryptography>=41.0.0

adbutils>=1.2.0
pure-python-adb==0.3.0.dev0
uiautomator2>=2.16.0
opencv-python>=4.7.0

python-dotenv>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
requests>=2.30.0
pillow>=10.0.0
loguru>=0.7.0
reportlab>=4.4.3

pytest>=7.3.1
pytest-cov>=4.1.0

black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
